backend/requirements.txt
requests
pycoingecko
plyer
pymongo
python-dotenv
dnspython
flask
pync
coinmarketcap